refactor: refactored some code
